Normal Unbiased Data Generating Model

An example data generating model to simulate effect sizes without publication bias.

Usage

# S3 method for class 'no_bias'
dgm(dgm_name, settings)

Arguments

dgm_name

DGM name (automatically passed)

settings

List containing

mean_effect

Mean effect

heterogeneity

Effect heterogeneity

n_studies

Number of effect size estimates

Value

Data frame with

yi

effect size

sei

standard error

Details

Sample sizes of individual effect size estimates are generated from a negative binomial distribution based on empirical sample size distribution presented in Appendix B of Maier et al. (2023)

References

Maier M, Bartoš F, Wagenmakers E (2023). “Robust Bayesian meta-analysis: Addressing publication bias with model-averaging.” Psychological Methods, 28(1), 107-122. doi:10.1037/met0000405 .

See also