Download Datasets of a DGM

This function downloads datasets of a specified Data-Generating Mechanism (DGM). All data are located at https://osf.io/exf3m/.

Usage

download_dgm_datasets(
  dgm_name,
  path = NULL,
  overwrite = FALSE,
  progress = TRUE,
  max_try = 10
)

download_dgm_results(
  dgm_name,
  path = NULL,
  overwrite = FALSE,
  progress = TRUE,
  max_try = 10
)

download_dgm_measures(
  dgm_name,
  path = NULL,
  overwrite = FALSE,
  progress = TRUE,
  max_try = 10
)

Arguments

dgm_name

Character string specifying the name of the DGM dataset to download.

path

Character string specifying the directory path where the datasets/results/measures should be saved. Defaults to the location specified via PublicationBiasBenchmark.get_option("simulation_directory"). The objects are stored in dgm_name/datasets, dgm_name/results, dgm_name/measures subfolders.

overwrite

Logical indicating whether to overwrite existing files. Defaults to FALSE, which means only missing files will be downloaded.

progress

Logical indicating whether to show progress downloading files. Defaults to TRUE.

max_try

Integet specifying how many times should the function attempt reconnecting to OSF upon failure.

Value

TRUE if the download was successful, otherwise an error is raised.

Examples

if (FALSE) { # \dontrun{
  download_dgm_datasets("no_bias")
} # }