Retrieve a Pre-Computed Results of a Method Applied to DGM

This function returns a pre-computed results of a given method at a specific repetition and condition from a dgm. The pre-computed results must be already stored locally. See download_dgm_results() function for more guidance.

Usage

retrieve_dgm_results(
  dgm_name,
  method = NULL,
  method_setting = "default",
  condition_id = NULL,
  repetition_id = NULL,
  path = NULL
)

Arguments

dgm_name

Character string specifying the DGM type

method

Which method should be returned. The complete results can be returned by setting to NULL.

method_setting

Which method setting should be returned. Defaults to "default".

condition_id

which conditions should settings be returned for.

repetition_id

Which repetition should be returned. The complete condition can be returned by setting to either NULL.

path

Character string specifying the directory path where the datasets/results/measures should be saved. Defaults to the location specified via PublicationBiasBenchmark.get_option("simulation_directory"). The objects are stored in dgm_name/datasets, dgm_name/results, dgm_name/measures subfolders.

Value

A data.frame

Examples

if (FALSE) { # \dontrun{
  # get condition 1, repetition 1 for default method setting
  retrieve_dgm_results("no_bias", condition_id = 1, repetition_id = 1)

  # get condition 1, all repetitions for default method setting
  retrieve_dgm_results("no_bias", condition_id = 1)
} # }