Simulate From Data Generating Mechanism

This function provides a unified interface to various data generating mechanisms for simulation studies. The specific DGM is determined by the first argument.

Usage

simulate_dgm(dgm_name, settings)

Arguments

dgm_name

Character string specifying the DGM type

settings

List containing the required parameters for the DGM or numeric condition_id

Value

A data frame containing the generated data with standardized structure

See also

Examples

simulate_dgm("Carter2019", list(mean_effect = 0, effect_heterogeneity = 0,
                       bias = "high", QRP = "high", n_studies = 10))
#>            yi        sei   ni
#> 1  0.10613077 0.05411249 1368
#> 2  0.66268909 0.31780816   42
#> 3  0.19643911 0.09378541  457
#> 4  0.62623920 0.27926695   54
#> 5  0.59698366 0.30225019   46
#> 6  0.52158709 0.25059508   66
#> 7  1.43281151 0.57774893   16
#> 8  1.16265101 0.55242535   16
#> 9  0.03548516 0.06788979  868
#> 10 0.57225009 0.28342330   52

simulate_dgm("Alinaghi2018", list(environment = "FE", mean_effect = 0))
#> Error in if (bias != "none") {    if (type == "PRE") {        if (bias == "significant") {            MetaData <- subset(MetaData, ((MetaData$nSig >= 0.7) |                 (MetaData$rnd < 0.1)))        }        else {            MetaData <- subset(MetaData, ((MetaData$nPos >= 0.7) |                 (MetaData$rnd < 0.1)))        }    }    else {        if (bias == "significant") {            MetaData <- subset(MetaData, ((MetaData$Sig == 1) |                 (MetaData$rnd < 0.1)))        }        else {            MetaData <- subset(MetaData, ((MetaData$effect >                 0) | (MetaData$rnd < 0.1)))        }    }}: argument is of length zero

simulate_dgm("Stanley2017", list(environment = "Cohens_d", mean_effect = 0,
                        effect_heterogeneity = 0, bias = 0, n_studies = 5,
                        sample_sizes = c(32,64,125,250,500)))
#>              yi        sei   ni
#> 1  0.1012553489 0.25016015   64
#> 2  0.0650198160 0.17682340  128
#> 3 -0.3028703590 0.12721423  250
#> 4  0.0297883251 0.08944768  500
#> 5  0.0008389488 0.06324556 1000