Simulate From Data Generating Mechanism

This function provides a unified interface to various data generating mechanisms for simulation studies. The specific DGM is determined by the first argument.

Usage

simulate_dgm(dgm_name, settings)

Arguments

dgm_name

Character string specifying the DGM type

settings

List containing the required parameters for the DGM or numeric condition_id

Value

A data frame containing the generated data with standardized structure

See also

Examples

simulate_dgm("Carter2019", list(mean_effect = 0, effect_heterogeneity = 0,
                       bias = "high", QRP = "high", n_studies = 10))
#>            yi        sei   ni
#> 1  0.10613077 0.05411249 1368
#> 2  0.66268909 0.31780816   42
#> 3  0.19643911 0.09378541  457
#> 4  0.62623920 0.27926695   54
#> 5  0.59698366 0.30225019   46
#> 6  0.52158709 0.25059508   66
#> 7  1.43281151 0.57774893   16
#> 8  1.16265101 0.55242535   16
#> 9  0.03548516 0.06788979  868
#> 10 0.57225009 0.28342330   52

simulate_dgm("Alinaghi2018", list(environment = "FE", mean_effect = 0))
#>                 yi        sei  ni study_id es_id
#> 1     1.4161757744 1.95133846 100        1     1
#> 2     0.9717106912 1.60326388 100        2     1
#> 3     0.7435443209 0.95640198 100        3     1
#> 4     0.2848260122 0.88101869 100        4     1
#> 5    -0.9683645053 2.14789003 100        5     1
#> 6    -0.2525241785 1.17673693 100        6     1
#> 7     4.1811297040 2.71442222 100        7     1
#> 8    -0.4345318981 0.55175254 100        8     1
#> 9    -0.2952125436 2.03009976 100        9     1
#> 10    0.1249059718 1.67418151 100       10     1
#> 11    2.3072081246 2.75672105 100       11     1
#> 12    1.5484428730 1.67681260 100       12     1
#> 13    0.1683050966 0.94466703 100       13     1
#> 14   -1.2753903035 1.39118089 100       14     1
#> 15    0.0824251096 0.48936299 100       15     1
#> 16   -3.0576105257 2.85017033 100       16     1
#> 17    0.5507368462 0.78208465 100       17     1
#> 18   -2.8107421547 1.71461703 100       18     1
#> 19    0.3223259113 0.40758119 100       19     1
#> 20   -0.6459930371 3.06228027 100       20     1
#> 21   -0.8306783526 1.56989342 100       21     1
#> 22    0.0470305549 1.02074183 100       22     1
#> 23   -2.4462193227 2.18419549 100       23     1
#> 24   -0.5008912252 0.74385677 100       24     1
#> 25   -3.1384541050 2.26482257 100       25     1
#> 26    0.6310374591 0.70656725 100       26     1
#> 27   -3.3376809660 2.64815448 100       27     1
#> 28   -1.3073142530 1.48722517 100       28     1
#> 29   -0.9270280838 0.60631874 100       29     1
#> 30   -0.0112611187 0.53579423 100       30     1
#> 31   -0.0069576090 0.76650613 100       31     1
#> 32    0.0201935820 0.08681561 100       32     1
#> 33   -0.9135244151 0.90358745 100       33     1
#> 34   -0.3353403949 1.37431207 100       34     1
#> 35   -0.0717012801 2.70710114 100       35     1
#> 36    0.7088651761 0.61604574 100       36     1
#> 37    0.0583500901 0.52386189 100       37     1
#> 38   -1.0767219235 1.78483550 100       38     1
#> 39   -2.0021424437 3.17893588 100       39     1
#> 40    1.9163192738 2.73506085 100       40     1
#> 41    0.9327232343 1.80871107 100       41     1
#> 42    1.6688577680 1.85254585 100       42     1
#> 43    2.4118215319 2.06556566 100       43     1
#> 44   -0.5507715729 2.71376572 100       44     1
#> 45    0.0417806255 1.39890744 100       45     1
#> 46   -1.5914577299 2.41418330 100       46     1
#> 47   -1.6555642851 1.64109181 100       47     1
#> 48   -0.5878310133 1.97380780 100       48     1
#> 49    0.1910179828 0.68999113 100       49     1
#> 50    2.8276665251 2.51007346 100       50     1
#> 51   -1.7261917931 2.04636145 100       51     1
#> 52   -0.1666557913 2.13404377 100       52     1
#> 53    0.8354672342 2.45675232 100       53     1
#> 54   -2.5102778034 2.09790882 100       54     1
#> 55   -0.9957992633 2.54006514 100       55     1
#> 56    5.8935937092 2.90163398 100       56     1
#> 57    1.4534294157 2.12905864 100       57     1
#> 58    0.3171971173 2.62025443 100       58     1
#> 59    0.0104060671 0.56731954 100       59     1
#> 60    0.4492158251 0.73584849 100       60     1
#> 61   -4.1154567751 2.66806187 100       61     1
#> 62    0.4425525330 0.67042547 100       62     1
#> 63    2.3827050464 1.62835699 100       63     1
#> 64    0.0309384151 0.27746603 100       64     1
#> 65   -0.4483438449 1.08299838 100       65     1
#> 66   -0.1198050674 0.52354566 100       66     1
#> 67    0.5974635293 0.73211035 100       67     1
#> 68   -0.3933917926 0.89923097 100       68     1
#> 69   -0.4845269964 1.68146859 100       69     1
#> 70    2.2719504061 2.61660116 100       70     1
#> 71   -0.9139121534 2.51363469 100       71     1
#> 72    1.1000528804 2.46647602 100       72     1
#> 73    0.9167345296 1.30058277 100       73     1
#> 74    1.9605230652 2.88150536 100       74     1
#> 75    0.7128683826 1.13050865 100       75     1
#> 76    4.0301575880 1.97257878 100       76     1
#> 77   -2.1099357642 1.87800860 100       77     1
#> 78   -0.3071160713 1.80068536 100       78     1
#> 79   -0.7579323734 1.53333249 100       79     1
#> 80    2.0820796412 1.71398981 100       80     1
#> 81   -0.2403435312 1.22454603 100       81     1
#> 82    0.3709272949 0.26066782 100       82     1
#> 83   -0.1041699964 0.51623134 100       83     1
#> 84    1.2830254651 1.37932260 100       84     1
#> 85   -1.5821294529 2.93441894 100       85     1
#> 86    5.9876813739 2.87003804 100       86     1
#> 87   -3.7956804329 2.77029579 100       87     1
#> 88   -1.6029108037 2.16383005 100       88     1
#> 89   -0.6933593800 0.79061767 100       89     1
#> 90    0.2622885214 0.89181432 100       90     1
#> 91   -0.0101653751 0.06976047 100       91     1
#> 92   -1.1248538974 1.16754519 100       92     1
#> 93    1.1687501972 0.59944757 100       93     1
#> 94   -0.6239910385 2.13316756 100       94     1
#> 95   -1.4176884247 2.79448333 100       95     1
#> 96    0.8186490675 1.26784140 100       96     1
#> 97   -0.2456288015 1.17039565 100       97     1
#> 98    1.5297306777 1.55015597 100       98     1
#> 99   -0.1506179654 1.38772950 100       99     1
#> 100   2.4038157537 2.75633687 100      100     1
#> 101  -0.0121314378 0.18626118 100      101     1
#> 102   1.2394989035 1.70353303 100      102     1
#> 103  -0.3590864952 0.43792101 100      103     1
#> 104  -0.9046058868 1.09866372 100      104     1
#> 105  -0.6350777369 2.24416920 100      105     1
#> 106   0.3508062381 0.43429196 100      106     1
#> 107   0.7685965827 1.22510197 100      107     1
#> 108   2.0774681524 2.43111861 100      108     1
#> 109  -0.9323012014 1.34160077 100      109     1
#> 110  -0.3284124203 1.39378086 100      110     1
#> 111   1.1826953817 1.28261249 100      111     1
#> 112  -0.8426076235 2.05552418 100      112     1
#> 113  -1.3343183432 3.03586546 100      113     1
#> 114  -1.4155467439 1.19995348 100      114     1
#> 115   0.4558202227 1.63976664 100      115     1
#> 116   2.1239236778 1.55382029 100      116     1
#> 117   0.1260249201 2.22213841 100      117     1
#> 118   3.3378612596 3.29720903 100      118     1
#> 119   6.2307799918 1.92774260 100      119     1
#> 120   2.4990369597 0.84253698 100      120     1
#> 121   0.0891739371 1.54076764 100      121     1
#> 122   4.3476690304 2.47450271 100      122     1
#> 123   1.8449064205 2.46705443 100      123     1
#> 124   2.0340427749 1.36980058 100      124     1
#> 125  -3.3749654126 2.74401057 100      125     1
#> 126   0.8028243473 2.00296878 100      126     1
#> 127  -0.3562869853 0.74369905 100      127     1
#> 128   0.0556246881 0.04001694 100      128     1
#> 129  -2.5815870778 1.87107555 100      129     1
#> 130   0.3114301341 0.31953878 100      130     1
#> 131   3.6443701572 2.83932868 100      131     1
#> 132  -2.7099587499 2.56607864 100      132     1
#> 133  -1.1179859676 1.50496415 100      133     1
#> 134  -0.0462755665 0.47290435 100      134     1
#> 135  -0.1319774074 0.35196896 100      135     1
#> 136  -1.5000078389 0.88138784 100      136     1
#> 137  -0.0776811401 0.92301371 100      137     1
#> 138   1.6445083485 0.72404544 100      138     1
#> 139   0.7052839258 1.49825115 100      139     1
#> 140   2.8082676724 2.27444579 100      140     1
#> 141   0.5294443620 0.51214700 100      141     1
#> 142   2.1308978221 2.39688829 100      142     1
#> 143   4.3985138014 2.89475345 100      143     1
#> 144   0.3756000450 1.17595643 100      144     1
#> 145  -3.4456178476 2.73450047 100      145     1
#> 146  -1.2896006885 1.13260795 100      146     1
#> 147   4.1999096080 1.84544988 100      147     1
#> 148  -1.8451856576 2.01882979 100      148     1
#> 149   0.7358662458 0.78195625 100      149     1
#> 150  -2.9548965152 2.64607368 100      150     1
#> 151   0.8983215900 1.63756783 100      151     1
#> 152  -0.1276006381 0.41036558 100      152     1
#> 153  -0.1651191647 0.40545319 100      153     1
#> 154  -2.1346178351 2.87472230 100      154     1
#> 155  -0.1589325890 0.33274834 100      155     1
#> 156  -0.0353067688 0.98545141 100      156     1
#> 157   3.8184517387 3.17200096 100      157     1
#> 158   1.5803536738 1.48349322 100      158     1
#> 159  -3.0430737535 2.14263470 100      159     1
#> 160   1.8073028356 2.34313268 100      160     1
#> 161   0.0809614392 0.82012631 100      161     1
#> 162   0.0903001743 0.74477483 100      162     1
#> 163   1.5910737210 3.50954504 100      163     1
#> 164   2.8270977182 2.38967349 100      164     1
#> 165  -0.1054459186 1.46489142 100      165     1
#> 166  -0.0103606900 2.48908535 100      166     1
#> 167   3.6866768050 1.80182099 100      167     1
#> 168  -0.1071011258 0.21708780 100      168     1
#> 169   1.0726899321 0.67415921 100      169     1
#> 170  -0.0638157574 1.02747668 100      170     1
#> 171   0.0287518854 0.39033731 100      171     1
#> 172   0.1110083456 0.07823031 100      172     1
#> 173   0.2673008738 1.47829570 100      173     1
#> 174  -1.6174536267 2.93809744 100      174     1
#> 175   0.2609677164 0.68796370 100      175     1
#> 176   0.3918490623 0.33093568 100      176     1
#> 177  -0.0904691283 0.35763461 100      177     1
#> 178  -0.6452823779 2.26580440 100      178     1
#> 179   3.3195131665 1.77273150 100      179     1
#> 180   1.0609237791 1.16912521 100      180     1
#> 181  -1.3739369210 1.25136282 100      181     1
#> 182  -0.0715568358 0.23437679 100      182     1
#> 183  -1.1271753012 2.00250532 100      183     1
#> 184  -0.4926964095 1.42505935 100      184     1
#> 185  -0.0298321867 2.28413865 100      185     1
#> 186   1.2096450660 1.48374270 100      186     1
#> 187  -3.1958097360 2.03492992 100      187     1
#> 188   4.2011386918 1.76204992 100      188     1
#> 189   4.1746785718 1.84950526 100      189     1
#> 190   0.0865316878 1.36091229 100      190     1
#> 191  -1.8607010940 2.55777789 100      191     1
#> 192   1.0950499350 2.13626248 100      192     1
#> 193  -0.0705862244 0.29754894 100      193     1
#> 194   3.4558547659 3.18021924 100      194     1
#> 195  -0.0154426957 0.21874667 100      195     1
#> 196  -3.2581212402 2.66982170 100      196     1
#> 197  -0.3846536612 0.34005489 100      197     1
#> 198  -1.1248807276 3.15691488 100      198     1
#> 199  -1.4558396401 1.32288846 100      199     1
#> 200  -0.0680315479 0.39578584 100      200     1
#> 201  -0.3102419791 0.56983919 100      201     1
#> 202   2.2756671276 2.59659420 100      202     1
#> 203  -3.4859004443 1.94710721 100      203     1
#> 204  -0.8544978467 2.74077076 100      204     1
#> 205  -0.0245684042 0.16441316 100      205     1
#> 206  -0.1708177192 1.24946852 100      206     1
#> 207  -0.1496858286 0.32602423 100      207     1
#> 208  -2.5242586425 1.96729192 100      208     1
#> 209   1.9277255407 1.27861562 100      209     1
#> 210   0.8810743478 1.79405294 100      210     1
#> 211  -1.9163064812 1.14521389 100      211     1
#> 212  -0.8469928738 0.97376667 100      212     1
#> 213   0.0467813974 0.15142362 100      213     1
#> 214   0.0039655225 0.04664660 100      214     1
#> 215   0.3645764212 1.76776785 100      215     1
#> 216   0.1438430774 1.88323810 100      216     1
#> 217   0.7763565269 1.96772307 100      217     1
#> 218  -0.3166372193 0.85641681 100      218     1
#> 219  -0.1135110821 0.43431382 100      219     1
#> 220  -0.4693214704 2.09138781 100      220     1
#> 221  -0.8373222166 1.15581176 100      221     1
#> 222  -0.9593407391 1.85397174 100      222     1
#> 223  -1.6568344801 2.26270790 100      223     1
#> 224   0.1039216942 0.37894066 100      224     1
#> 225  -6.9239543841 2.24564830 100      225     1
#> 226   0.8568797174 1.01739916 100      226     1
#> 227  -0.5127721804 1.28068545 100      227     1
#> 228   0.0477812235 0.31716272 100      228     1
#> 229   0.3216788267 1.55295828 100      229     1
#> 230   2.3567641934 1.61624939 100      230     1
#> 231   0.7452774155 1.22434857 100      231     1
#> 232  -0.1066480401 0.57080180 100      232     1
#> 233  -2.8709920996 1.52510400 100      233     1
#> 234  -0.7101333589 1.00229258 100      234     1
#> 235   2.2141640132 1.44347561 100      235     1
#> 236   1.2307181053 1.08973097 100      236     1
#> 237  -0.4506383206 0.37759418 100      237     1
#> 238   1.1678375669 2.27874624 100      238     1
#> 239   0.1585466932 0.60938481 100      239     1
#> 240   0.0337890955 0.17200108 100      240     1
#> 241  -1.2223945657 2.29077480 100      241     1
#> 242  -1.1069314632 2.97723223 100      242     1
#> 243  -1.2029288762 1.09368572 100      243     1
#> 244   0.3853956594 0.54168822 100      244     1
#> 245   5.5118432235 3.37932510 100      245     1
#> 246  -0.2823047159 0.28989763 100      246     1
#> 247  -0.4166909128 0.52631634 100      247     1
#> 248   0.7524974778 1.89946075 100      248     1
#> 249  -1.2016797975 1.72123309 100      249     1
#> 250   0.2758725948 2.44088920 100      250     1
#> 251   0.5198136965 2.20429989 100      251     1
#> 252  -0.8841546970 0.33098861 100      252     1
#> 253  -2.0599032836 1.27496135 100      253     1
#> 254   1.7464476009 2.72965350 100      254     1
#> 255  -0.2085958235 0.81178530 100      255     1
#> 256  -4.3459285489 2.15502109 100      256     1
#> 257  -2.2562613948 1.57181406 100      257     1
#> 258  -2.8970208958 2.07588419 100      258     1
#> 259   0.2360316013 3.45579033 100      259     1
#> 260  -1.2447281819 3.21124321 100      260     1
#> 261  -0.2774354241 1.17222273 100      261     1
#> 262  -1.2886313605 1.68220256 100      262     1
#> 263  -2.0152348006 1.23032009 100      263     1
#> 264   0.2775016647 1.54944935 100      264     1
#> 265  -2.0010600767 0.92639124 100      265     1
#> 266  -0.7779363844 0.91526094 100      266     1
#> 267   0.8599811263 1.18219994 100      267     1
#> 268   2.4083003038 1.37448441 100      268     1
#> 269  -0.6249285177 1.02033411 100      269     1
#> 270   0.6927537983 3.54745145 100      270     1
#> 271   3.5231681058 2.88367142 100      271     1
#> 272   0.8070936185 1.10656472 100      272     1
#> 273   1.7683455332 1.95606748 100      273     1
#> 274   2.9879000756 2.07937340 100      274     1
#> 275  -5.2764736799 2.22394720 100      275     1
#> 276  -0.6547093934 2.53705296 100      276     1
#> 277  -0.5912392372 2.03273555 100      277     1
#> 278   3.0540167954 1.46055135 100      278     1
#> 279  -1.3331459068 2.62118133 100      279     1
#> 280   0.2681012202 0.33177892 100      280     1
#> 281   3.6782507509 1.29955135 100      281     1
#> 282  -0.9961354919 2.19585199 100      282     1
#> 283  -0.2132937374 0.37715215 100      283     1
#> 284   0.2527093478 0.33055630 100      284     1
#> 285   1.0273174814 1.89218429 100      285     1
#> 286  -0.1725348083 0.81261498 100      286     1
#> 287   1.4272537161 2.49990040 100      287     1
#> 288   3.4845760921 1.42030666 100      288     1
#> 289  -2.9461073304 3.16745011 100      289     1
#> 290   0.0031573190 0.41078680 100      290     1
#> 291   0.0159543360 2.06443807 100      291     1
#> 292   0.1836892692 0.38827324 100      292     1
#> 293   2.6215167914 2.85931041 100      293     1
#> 294   0.7191997177 0.59271832 100      294     1
#> 295   0.1268138271 0.08414269 100      295     1
#> 296   0.2239423510 0.09357793 100      296     1
#> 297   1.1365812051 0.79212277 100      297     1
#> 298  -0.2433951019 0.56148334 100      298     1
#> 299   1.0572026850 0.71046873 100      299     1
#> 300   0.6204893356 0.40623452 100      300     1
#> 301  -0.5468577360 0.69027200 100      301     1
#> 302  -0.9127332500 1.68403514 100      302     1
#> 303  -0.2911178280 1.23726085 100      303     1
#> 304   2.1973912706 1.81542710 100      304     1
#> 305  -0.1389935414 2.01055512 100      305     1
#> 306  -0.5364525670 1.07097656 100      306     1
#> 307   0.7349719829 1.46180562 100      307     1
#> 308  -2.6566343457 3.14744205 100      308     1
#> 309   0.0142138880 0.68724178 100      309     1
#> 310  -2.6939550445 2.09098701 100      310     1
#> 311   0.0691290666 2.57823275 100      311     1
#> 312   0.6283479289 3.38183117 100      312     1
#> 313  -2.2701025083 1.67976750 100      313     1
#> 314  -6.5236959382 3.04663944 100      314     1
#> 315   2.1097138091 2.29354234 100      315     1
#> 316   1.8175721044 0.83305636 100      316     1
#> 317   1.3458975271 1.03481097 100      317     1
#> 318   2.2888883139 2.84957827 100      318     1
#> 319   1.0371295395 1.08267159 100      319     1
#> 320  -0.3733412938 2.01109090 100      320     1
#> 321   0.6591220098 0.81578424 100      321     1
#> 322   1.9923116120 2.80590752 100      322     1
#> 323  -0.3063022795 0.33098720 100      323     1
#> 324  -0.0422227954 0.38394229 100      324     1
#> 325  -0.3400939019 0.40142768 100      325     1
#> 326  -1.9511296436 2.38058286 100      326     1
#> 327   0.5373260133 0.41336591 100      327     1
#> 328   0.1306211280 1.77746487 100      328     1
#> 329   2.1967619875 2.18685226 100      329     1
#> 330   0.0694879298 0.10196735 100      330     1
#> 331  -1.3214836930 1.53856031 100      331     1
#> 332   1.4942412926 1.38766798 100      332     1
#> 333   0.8776060422 0.81961045 100      333     1
#> 334  -0.2799503361 1.01307046 100      334     1
#> 335   0.4643448679 1.52143740 100      335     1
#> 336   0.5815865948 1.39897619 100      336     1
#> 337  -2.7730521470 2.22795768 100      337     1
#> 338  -1.0150963731 0.71710253 100      338     1
#> 339   0.1017349457 0.32196225 100      339     1
#> 340  -1.0242897104 0.55185546 100      340     1
#> 341  -0.7855536493 0.67846362 100      341     1
#> 342  -2.3638058921 2.24203849 100      342     1
#> 343  -0.1125678206 0.48037680 100      343     1
#> 344   0.0844713955 2.33307117 100      344     1
#> 345   0.0233444810 1.48178420 100      345     1
#> 346  -2.4976499958 1.88052830 100      346     1
#> 347  -0.4850927980 1.73389817 100      347     1
#> 348   1.0449813617 0.74700751 100      348     1
#> 349  -0.4571741858 1.17932098 100      349     1
#> 350  -0.1396979416 1.00148625 100      350     1
#> 351   1.6826001708 2.55523436 100      351     1
#> 352   1.7042567251 2.89002897 100      352     1
#> 353   4.0695198349 2.25997496 100      353     1
#> 354   0.6580538488 2.41738100 100      354     1
#> 355   0.3306264276 2.75421644 100      355     1
#> 356   1.8693772131 1.62508846 100      356     1
#> 357  -1.7455009730 3.28430545 100      357     1
#> 358  -1.2793783065 0.67654500 100      358     1
#> 359  -0.0195848409 0.32038724 100      359     1
#> 360   2.7102992088 2.63347252 100      360     1
#> 361   1.6617006766 3.08439322 100      361     1
#> 362   2.8734388370 2.12292159 100      362     1
#> 363   1.9203498293 1.93397790 100      363     1
#> 364   0.8074730611 1.32363695 100      364     1
#> 365   0.9211196950 0.86529106 100      365     1
#> 366  -1.1469643697 1.38587506 100      366     1
#> 367   0.4414623110 0.58742493 100      367     1
#> 368  -1.2099997783 2.05018166 100      368     1
#> 369   2.7151500344 2.38776955 100      369     1
#> 370   2.6173734570 3.03912454 100      370     1
#> 371  -6.4451030558 2.85094618 100      371     1
#> 372  -2.7346431873 2.73945854 100      372     1
#> 373   0.4553688907 0.57731212 100      373     1
#> 374   0.1168473484 0.06520175 100      374     1
#> 375  -0.4725049191 0.26658851 100      375     1
#> 376  -1.6514295552 2.83767014 100      376     1
#> 377   2.3084338548 1.60091624 100      377     1
#> 378   0.4808162590 2.87430636 100      378     1
#> 379   0.5115687167 2.73653235 100      379     1
#> 380  -4.3479202663 3.19641862 100      380     1
#> 381  -0.1234680229 0.39573839 100      381     1
#> 382  -0.7147013048 3.16562965 100      382     1
#> 383   0.3197032566 2.19996532 100      383     1
#> 384   0.2061423029 0.27531746 100      384     1
#> 385   0.0887176759 0.09789545 100      385     1
#> 386  -0.0571369389 0.04724818 100      386     1
#> 387  -0.0280220192 0.40788831 100      387     1
#> 388   0.3636188711 0.50540371 100      388     1
#> 389  -1.2641983101 1.98869845 100      389     1
#> 390   0.3021486232 0.12783636 100      390     1
#> 391   1.1029255885 2.88992505 100      391     1
#> 392  -3.2415355759 2.54382600 100      392     1
#> 393   1.2431068606 2.70472995 100      393     1
#> 394  -0.6366590033 3.15110986 100      394     1
#> 395  -0.8806765529 2.06366663 100      395     1
#> 396  -0.0482437921 0.16736494 100      396     1
#> 397  -1.4027235699 2.38946600 100      397     1
#> 398  -0.2268859715 0.18378854 100      398     1
#> 399  -0.3794715358 1.80163262 100      399     1
#> 400  -1.6935900326 0.78854304 100      400     1
#> 401   0.0193045085 0.31946540 100      401     1
#> 402  -2.1474700847 2.27862372 100      402     1
#> 403   0.2678417776 1.56208344 100      403     1
#> 404   3.1314640778 1.96817227 100      404     1
#> 405  -1.8799618119 2.32547211 100      405     1
#> 406   3.7782365676 2.42387625 100      406     1
#> 407   0.7803012988 1.82890596 100      407     1
#> 408  -0.4392462085 1.16111242 100      408     1
#> 409  -1.4540558198 1.61434524 100      409     1
#> 410  -1.7709372313 0.78817882 100      410     1
#> 411  -0.3170845342 0.28595017 100      411     1
#> 412  -0.8115708454 1.71166267 100      412     1
#> 413  -0.7335172113 1.99359032 100      413     1
#> 414   1.5094402392 2.02036060 100      414     1
#> 415  -4.6731412809 2.94312600 100      415     1
#> 416  -2.3220184692 1.17088919 100      416     1
#> 417  -2.1603551906 2.85213301 100      417     1
#> 418   5.1269501756 2.44521747 100      418     1
#> 419  -2.6106300432 2.60932166 100      419     1
#> 420   0.0008299109 0.14065285 100      420     1
#> 421   0.4534224525 1.01538252 100      421     1
#> 422  -0.0408095543 0.16293163 100      422     1
#> 423  -1.7852032467 2.51443745 100      423     1
#> 424   0.0733029555 0.32630002 100      424     1
#> 425   3.8930322453 2.47858026 100      425     1
#> 426  -0.5262284887 0.69202701 100      426     1
#> 427  -0.4645413776 1.82056894 100      427     1
#> 428   0.4828651738 1.65269965 100      428     1
#> 429  -0.4569786555 2.56437255 100      429     1
#> 430  -0.2131952370 0.25930232 100      430     1
#> 431  -0.0035414755 0.02267323 100      431     1
#> 432  -3.9800888821 2.93516597 100      432     1
#> 433  -1.8181311896 1.76772568 100      433     1
#> 434  -0.2901460844 0.84260724 100      434     1
#> 435   2.1933872641 0.96088050 100      435     1
#> 436  -0.0585467617 0.06300586 100      436     1
#> 437  -0.3406273566 2.86666079 100      437     1
#> 438   0.9773397337 3.11017494 100      438     1
#> 439  -0.4125661017 0.90406782 100      439     1
#> 440  -0.0116393628 0.31404588 100      440     1
#> 441   0.4677304442 0.85422606 100      441     1
#> 442  -0.2205046325 0.13281542 100      442     1
#> 443   1.5341496140 2.39334508 100      443     1
#> 444   0.0024380736 0.24056217 100      444     1
#> 445  -0.1851431685 0.17314821 100      445     1
#> 446  -0.7178383353 1.61999671 100      446     1
#> 447  -0.3622696220 0.26540350 100      447     1
#> 448  -0.3425341575 0.67500011 100      448     1
#> 449  -1.5290416872 1.64396713 100      449     1
#> 450   1.8546362330 1.83047402 100      450     1
#> 451   4.9375175675 2.26197751 100      451     1
#> 452  -0.1537760683 1.55427640 100      452     1
#> 453  -2.8319517704 1.67614042 100      453     1
#> 454   2.4756337466 2.61983436 100      454     1
#> 455  -1.5916399688 2.69924185 100      455     1
#> 456  -1.6140534254 0.55415835 100      456     1
#> 457  -0.0207188150 0.07893822 100      457     1
#> 458  -0.2716746519 1.74411860 100      458     1
#> 459  -0.3374812731 1.14534022 100      459     1
#> 460  -0.1031377058 0.22199562 100      460     1
#> 461  -1.7470136922 1.42203452 100      461     1
#> 462   1.4432108990 1.43095416 100      462     1
#> 463  -0.2764655476 0.27634943 100      463     1
#> 464  -0.0411621155 0.76653241 100      464     1
#> 465   2.1182148403 2.13573507 100      465     1
#> 466   0.0622744723 2.27323343 100      466     1
#> 467  -0.1273274936 1.54219010 100      467     1
#> 468  -1.0127398036 1.08956291 100      468     1
#> 469   1.7597192787 1.64892043 100      469     1
#> 470   0.1518796382 0.11514559 100      470     1
#> 471  -1.0196454368 1.74376644 100      471     1
#> 472  -0.6911796946 2.61060900 100      472     1
#> 473  -0.2151590352 1.89391782 100      473     1
#> 474   1.0808406361 1.53645165 100      474     1
#> 475  -0.2831464117 2.99569123 100      475     1
#> 476   3.2537382651 2.19217941 100      476     1
#> 477  -0.2185404852 2.72060123 100      477     1
#> 478  -0.6856241603 0.66945479 100      478     1
#> 479  -2.2983205898 2.73385254 100      479     1
#> 480  -0.9629872515 1.73519210 100      480     1
#> 481  -0.2194579062 1.29627099 100      481     1
#> 482   0.6342703641 2.06393570 100      482     1
#> 483   1.9296271808 1.62703167 100      483     1
#> 484   0.6945900233 1.43522866 100      484     1
#> 485   2.1670872576 2.11551568 100      485     1
#> 486  -0.8804522978 1.89677874 100      486     1
#> 487  -0.7166183282 0.68759937 100      487     1
#> 488   3.1691595149 2.73303072 100      488     1
#> 489  -0.3993221122 1.32157130 100      489     1
#> 490   0.0144875067 0.12339002 100      490     1
#> 491   0.1862350224 0.16884624 100      491     1
#> 492   2.1187871839 2.32623272 100      492     1
#> 493   2.6899854098 1.94216725 100      493     1
#> 494   0.5503552864 0.88686898 100      494     1
#> 495  -0.0349737226 0.06652110 100      495     1
#> 496   1.4393630563 2.30161499 100      496     1
#> 497   0.0603145899 0.85854516 100      497     1
#> 498   0.5265758945 2.33684863 100      498     1
#> 499   0.6730001950 0.52362928 100      499     1
#> 500  -1.1129996295 3.36722383 100      500     1
#> 501  -5.4914626749 2.97088432 100      501     1
#> 502   1.3477706463 3.48872861 100      502     1
#> 503   0.3826957079 2.00728211 100      503     1
#> 504  -0.0482345824 0.18035955 100      504     1
#> 505  -0.1521899769 1.62599900 100      505     1
#> 506  -0.9487504617 0.72228103 100      506     1
#> 507   1.0794536093 0.64348695 100      507     1
#> 508  -0.0052331014 0.67108256 100      508     1
#> 509  -0.0605622379 0.14932614 100      509     1
#> 510  -0.9762743522 0.67195587 100      510     1
#> 511   0.4306169665 2.45125779 100      511     1
#> 512  -2.9287432268 2.20459909 100      512     1
#> 513  -0.1218753944 1.14610148 100      513     1
#> 514   0.5961660106 0.79113368 100      514     1
#> 515  -0.1955379152 0.24977551 100      515     1
#> 516   0.3139123138 2.23888115 100      516     1
#> 517  -0.1155509701 2.13291658 100      517     1
#> 518   0.2373285072 1.96796738 100      518     1
#> 519  -0.1363255956 0.08571192 100      519     1
#> 520  -0.9070638576 1.41624139 100      520     1
#> 521  -1.9198460210 2.46002674 100      521     1
#> 522   3.3793200062 1.67100400 100      522     1
#> 523   0.1465415449 0.13290412 100      523     1
#> 524  -1.5743141201 2.92884144 100      524     1
#> 525  -0.4584254998 0.74974488 100      525     1
#> 526   1.0127705021 1.81883439 100      526     1
#> 527   0.7509107678 0.53476471 100      527     1
#> 528   0.5746934554 1.26305397 100      528     1
#> 529  -2.0489567224 2.49337473 100      529     1
#> 530  -0.3424044619 1.01835947 100      530     1
#> 531   0.6935811892 2.20791968 100      531     1
#> 532  -2.3049803831 3.10419791 100      532     1
#> 533  -0.9669929211 2.22322404 100      533     1
#> 534  -2.2987120192 2.10146990 100      534     1
#> 535  -0.1439720415 0.16186798 100      535     1
#> 536   3.2507085523 3.01818357 100      536     1
#> 537  -0.9718026814 2.33432936 100      537     1
#> 538   1.1813842778 2.98615275 100      538     1
#> 539   1.9015416910 1.61371224 100      539     1
#> 540  -0.7591666530 0.84832276 100      540     1
#> 541  -0.3368982629 0.59840538 100      541     1
#> 542   0.3680636580 1.64858744 100      542     1
#> 543   0.1633968991 2.81456595 100      543     1
#> 544   1.4519257458 1.74996090 100      544     1
#> 545  -1.9297548476 1.61976751 100      545     1
#> 546   0.0224646566 1.40347872 100      546     1
#> 547  -0.4083389696 0.31084756 100      547     1
#> 548  -0.8126948303 2.41573525 100      548     1
#> 549   0.4876853788 2.83165989 100      549     1
#> 550  -0.3785866227 1.75827305 100      550     1
#> 551  -0.7680957419 1.18976317 100      551     1
#> 552  -0.2826494965 0.73633696 100      552     1
#> 553  -0.7624712149 0.90268158 100      553     1
#> 554  -2.4740443346 2.45167965 100      554     1
#> 555  -0.4081552068 1.63158164 100      555     1
#> 556   0.8970264984 2.41134279 100      556     1
#> 557  -0.3828919482 0.71167566 100      557     1
#> 558   3.7495300378 2.80499091 100      558     1
#> 559  -0.6922389514 0.97920565 100      559     1
#> 560  -0.4009624000 0.59182431 100      560     1
#> 561  -3.1236800815 2.25584394 100      561     1
#> 562  -1.2794216473 0.92161425 100      562     1
#> 563   0.0514383629 0.18523860 100      563     1
#> 564   0.6946780501 1.35314185 100      564     1
#> 565  -2.4567496358 2.78109001 100      565     1
#> 566   0.0493729518 0.80483801 100      566     1
#> 567   0.6243145755 2.30092165 100      567     1
#> 568  -1.0589655303 2.06492734 100      568     1
#> 569   0.5094557715 1.16009256 100      569     1
#> 570  -0.8295206947 1.42602879 100      570     1
#> 571   0.5418714971 2.46666984 100      571     1
#> 572  -1.4066670035 2.70517299 100      572     1
#> 573  -4.5703859104 2.66996552 100      573     1
#> 574  -3.5157170446 2.94497488 100      574     1
#> 575   1.4703750152 2.77748789 100      575     1
#> 576   0.7184826026 0.74219849 100      576     1
#> 577   0.6622974843 2.05873958 100      577     1
#> 578  -1.6996423117 2.90668830 100      578     1
#> 579  -0.1355719612 0.19674328 100      579     1
#> 580   2.8848899413 1.95864620 100      580     1
#> 581   3.0913770598 2.82658163 100      581     1
#> 582  -1.8789969355 1.70395198 100      582     1
#> 583  -0.6769233500 2.01133207 100      583     1
#> 584   2.1244819745 2.01095557 100      584     1
#> 585  -0.7626529619 0.64736220 100      585     1
#> 586  -4.0968546112 2.49038628 100      586     1
#> 587   0.6234074812 0.40533625 100      587     1
#> 588  -0.1590546024 0.05455143 100      588     1
#> 589  -0.9018480712 1.98688592 100      589     1
#> 590   0.0995699089 0.80686766 100      590     1
#> 591   0.3809519515 0.78970188 100      591     1
#> 592   1.8029632573 1.65567270 100      592     1
#> 593   0.3517856264 0.41218644 100      593     1
#> 594   0.6244425042 0.26720569 100      594     1
#> 595  -0.1103443956 1.03643379 100      595     1
#> 596   0.0175238261 0.32335677 100      596     1
#> 597   2.5207747408 1.96677247 100      597     1
#> 598   0.3495139209 0.42511166 100      598     1
#> 599  -0.7677457841 2.56327892 100      599     1
#> 600  -0.0460153498 0.30470158 100      600     1
#> 601   0.8366243847 0.50567233 100      601     1
#> 602   0.0954857144 0.51752687 100      602     1
#> 603   0.7849100563 1.99147082 100      603     1
#> 604  -0.0858888974 0.85110038 100      604     1
#> 605  -1.7198471967 1.79088495 100      605     1
#> 606   0.1123503177 0.45689831 100      606     1
#> 607  -0.0413442242 0.42111911 100      607     1
#> 608  -1.3811988340 3.30858144 100      608     1
#> 609   3.4423021859 3.28822568 100      609     1
#> 610   0.0440258026 0.18772059 100      610     1
#> 611  -0.0886259519 0.76410781 100      611     1
#> 612   1.7114956982 1.82171884 100      612     1
#> 613   3.3024390506 1.90558400 100      613     1
#> 614   0.0825970167 0.54990988 100      614     1
#> 615  -1.8024734938 3.16971433 100      615     1
#> 616  -0.9717438002 1.32021925 100      616     1
#> 617   0.1734640455 1.07745904 100      617     1
#> 618   1.8943967654 2.64143794 100      618     1
#> 619   1.1924993280 0.81136260 100      619     1
#> 620  -0.9833151902 2.03727277 100      620     1
#> 621  -1.9618288866 1.59808383 100      621     1
#> 622   2.0148675205 1.64435647 100      622     1
#> 623  -0.4693313175 1.82426859 100      623     1
#> 624  -0.1153082544 0.24569050 100      624     1
#> 625  -2.5095276464 1.61397184 100      625     1
#> 626  -1.6157508750 2.29136347 100      626     1
#> 627   1.2600589069 1.24772407 100      627     1
#> 628   2.1838384232 2.18083131 100      628     1
#> 629  -1.8786481091 1.68864772 100      629     1
#> 630  -0.2847665522 1.46018814 100      630     1
#> 631   2.7342853788 1.13082314 100      631     1
#> 632   3.8751994129 2.72485637 100      632     1
#> 633   1.3266878268 1.36571036 100      633     1
#> 634  -1.6981843352 2.13188159 100      634     1
#> 635   0.1997730571 0.67589591 100      635     1
#> 636  -0.4614678776 0.75640765 100      636     1
#> 637   1.1548962618 2.73492824 100      637     1
#> 638   0.7367611096 1.55141591 100      638     1
#> 639   0.7392877944 0.80007963 100      639     1
#> 640   0.4134539682 2.50262959 100      640     1
#> 641  -1.0277922304 0.62639974 100      641     1
#> 642   0.3809105821 2.29485102 100      642     1
#> 643   2.2023339806 0.99247415 100      643     1
#> 644  -0.4309262298 1.29963622 100      644     1
#> 645   0.0529336744 1.75396642 100      645     1
#> 646  -0.9401279436 1.65840759 100      646     1
#> 647  -2.2345238825 2.61852520 100      647     1
#> 648   1.6080684372 1.82527204 100      648     1
#> 649   0.4208036708 2.16645959 100      649     1
#> 650   2.8418760723 1.95326892 100      650     1
#> 651  -1.7906415271 2.02260831 100      651     1
#> 652  -0.3586444535 0.75570435 100      652     1
#> 653   0.4602143443 2.92467391 100      653     1
#> 654  -0.4241177086 0.32332481 100      654     1
#> 655   1.2731056810 1.98532414 100      655     1
#> 656  -0.8942050576 0.38290984 100      656     1
#> 657  -2.0787642376 1.78283016 100      657     1
#> 658   2.0273445773 2.33640439 100      658     1
#> 659  -2.9922590408 2.49969397 100      659     1
#> 660   0.5882585865 0.93411204 100      660     1
#> 661  -1.1969187802 1.12253087 100      661     1
#> 662  -0.7915375053 0.84397848 100      662     1
#> 663  -3.6584814696 1.84959779 100      663     1
#> 664  -0.6513755871 1.57682618 100      664     1
#> 665  -4.4696504969 2.31515993 100      665     1
#> 666  -0.4350836121 1.39890185 100      666     1
#> 667   1.1462055400 2.42086194 100      667     1
#> 668   2.2401014783 2.46991654 100      668     1
#> 669   0.1418046867 0.17436841 100      669     1
#> 670  -1.3446409047 1.13966589 100      670     1
#> 671   1.0535639560 2.72618286 100      671     1
#> 672  -1.3234114733 1.62913226 100      672     1
#> 673  -0.9840359697 0.58041622 100      673     1
#> 674  -2.1136757154 2.78890646 100      674     1
#> 675  -0.1195637741 0.27094325 100      675     1
#> 676   0.2879302847 0.85063151 100      676     1
#> 677   3.2183499803 1.92984489 100      677     1
#> 678  -0.0171155622 0.02528572 100      678     1
#> 679   0.2845066566 0.41068251 100      679     1
#> 680  -0.0718734098 0.07912962 100      680     1
#> 681  -1.1755456885 1.93300906 100      681     1
#> 682   4.0536521749 3.14984034 100      682     1
#> 683  -0.1930923269 2.76826667 100      683     1
#> 684  -0.4904499282 2.77503096 100      684     1
#> 685   3.4738369315 2.55349623 100      685     1
#> 686  -1.7837703383 2.36203657 100      686     1
#> 687   1.2610471791 2.50069517 100      687     1
#> 688  -1.4510113681 1.19526601 100      688     1
#> 689  -0.0770565853 1.29262699 100      689     1
#> 690   0.7908381648 0.72471278 100      690     1
#> 691  -0.6263089333 0.64071738 100      691     1
#> 692   3.7620223498 2.17648877 100      692     1
#> 693  -0.7121982324 2.20953410 100      693     1
#> 694   1.0465378166 1.53173910 100      694     1
#> 695   1.9479668414 1.56365714 100      695     1
#> 696  -1.3273233054 1.84815945 100      696     1
#> 697  -0.5225497387 2.70217128 100      697     1
#> 698   1.0806894705 0.86524582 100      698     1
#> 699  -0.7657004581 0.84113403 100      699     1
#> 700  -2.0469782772 2.06480522 100      700     1
#> 701  -0.0373957874 0.03103023 100      701     1
#> 702   2.4976209943 1.17647308 100      702     1
#> 703   1.7785267323 1.41328792 100      703     1
#> 704   1.0343338524 0.48371280 100      704     1
#> 705   1.3274735836 2.45203389 100      705     1
#> 706  -2.0610583081 2.40368123 100      706     1
#> 707  -2.0099262641 2.89997250 100      707     1
#> 708  -1.4954150913 1.77490343 100      708     1
#> 709  -0.0367101415 1.61132564 100      709     1
#> 710  -2.0834914342 2.10730584 100      710     1
#> 711  -3.1874790767 2.41182730 100      711     1
#> 712   2.6043129993 1.55730347 100      712     1
#> 713   0.4674210199 1.95140698 100      713     1
#> 714   0.1637702013 1.05182845 100      714     1
#> 715  -0.1866282626 0.54951907 100      715     1
#> 716   2.2501913194 2.75573660 100      716     1
#> 717  -0.6677315878 1.22660805 100      717     1
#> 718   6.6153583678 3.31558490 100      718     1
#> 719  -1.0987634818 1.85743042 100      719     1
#> 720  -0.0701790637 0.45783072 100      720     1
#> 721  -0.6096415083 0.46998141 100      721     1
#> 722   0.6744588023 1.24494793 100      722     1
#> 723  -3.3462373958 2.61353850 100      723     1
#> 724  -0.1221889632 0.52124244 100      724     1
#> 725  -1.3626615458 1.44254977 100      725     1
#> 726   1.3544597236 1.51410880 100      726     1
#> 727  -6.2006794095 3.31896714 100      727     1
#> 728   0.7476348467 1.31785884 100      728     1
#> 729  -0.3571662718 1.76417616 100      729     1
#> 730  -0.1946109915 0.47321915 100      730     1
#> 731  -0.2206113427 0.33824224 100      731     1
#> 732   1.7403390982 2.86829688 100      732     1
#> 733  -0.7962469986 1.36447012 100      733     1
#> 734   0.4660500553 0.47712352 100      734     1
#> 735  -0.3320299855 0.66166752 100      735     1
#> 736  -0.7359498383 2.99609074 100      736     1
#> 737   3.1613055346 2.78325916 100      737     1
#> 738  -0.3372914536 1.73636749 100      738     1
#> 739  -1.6344195076 2.31825588 100      739     1
#> 740   2.0186453618 1.43241361 100      740     1
#> 741   1.7683205964 2.53932717 100      741     1
#> 742   1.6494835971 1.90767783 100      742     1
#> 743   0.8261590578 2.56436296 100      743     1
#> 744   0.1742161597 2.57585059 100      744     1
#> 745   1.7491955225 1.15266234 100      745     1
#> 746  -4.5168565506 2.35974255 100      746     1
#> 747  -2.1891796951 1.61047463 100      747     1
#> 748   0.7767891588 2.96283400 100      748     1
#> 749  -0.1513602054 0.23376862 100      749     1
#> 750   0.2951910211 0.14838000 100      750     1
#> 751  -0.2516387117 0.70627123 100      751     1
#> 752   0.8037243418 1.68075786 100      752     1
#> 753   0.3698394487 0.83867993 100      753     1
#> 754   0.0462943156 2.87787893 100      754     1
#> 755  -0.4777339376 1.09342685 100      755     1
#> 756   0.2847656120 0.26957893 100      756     1
#> 757   0.0167049416 0.11049551 100      757     1
#> 758   2.2290157132 1.96437838 100      758     1
#> 759  -1.0682024694 2.04494503 100      759     1
#> 760  -2.8114596769 2.88194659 100      760     1
#> 761   3.9059665776 2.46203121 100      761     1
#> 762  -1.8533944270 1.06139527 100      762     1
#> 763  -1.0131320632 0.97009603 100      763     1
#> 764   2.5062466104 2.64733329 100      764     1
#> 765   0.6062333940 0.57785773 100      765     1
#> 766   0.0213722263 0.34663551 100      766     1
#> 767   1.2426127329 0.97775756 100      767     1
#> 768   1.6247295060 2.96515141 100      768     1
#> 769  -0.5628540933 1.99928868 100      769     1
#> 770   0.0876967879 0.24134926 100      770     1
#> 771   1.1363979692 1.51371883 100      771     1
#> 772  -0.4245795220 0.84149638 100      772     1
#> 773  -4.0081678026 2.80159091 100      773     1
#> 774   3.9189588746 2.72332058 100      774     1
#> 775  -0.4698302187 0.54164748 100      775     1
#> 776  -0.4525283215 0.96723592 100      776     1
#> 777   0.4357610686 1.89014909 100      777     1
#> 778   1.9207694404 1.57725087 100      778     1
#> 779   0.8679387448 2.23357558 100      779     1
#> 780  -0.3153826389 2.21415191 100      780     1
#> 781   0.0733690983 2.04625487 100      781     1
#> 782  -1.9717740521 1.55709307 100      782     1
#> 783  -0.5783734926 0.52422179 100      783     1
#> 784  -0.1035193783 0.21935636 100      784     1
#> 785  -0.6007490794 0.52665674 100      785     1
#> 786  -2.1361442889 1.00720036 100      786     1
#> 787  -0.2342795347 1.87008256 100      787     1
#> 788  -0.7807172543 1.62453551 100      788     1
#> 789   1.8018521381 2.03989615 100      789     1
#> 790   0.1299642859 0.16221697 100      790     1
#> 791  -3.4868415708 2.65527517 100      791     1
#> 792   0.1811138816 0.98699930 100      792     1
#> 793   3.1265489596 2.90674598 100      793     1
#> 794   0.6419146262 0.62564486 100      794     1
#> 795   1.5244961402 3.20406478 100      795     1
#> 796  -0.4464551184 1.45792997 100      796     1
#> 797   1.6344328042 3.12127104 100      797     1
#> 798  -0.4431987648 2.09988453 100      798     1
#> 799  -0.3296871740 1.78147728 100      799     1
#> 800   0.1357351691 0.62703162 100      800     1
#> 801   0.9843215328 1.71010567 100      801     1
#> 802   3.4581530839 2.77901752 100      802     1
#> 803   0.4800888961 2.12076957 100      803     1
#> 804   1.3395757672 3.13477758 100      804     1
#> 805  -1.8402188718 1.56794203 100      805     1
#> 806   0.0989591147 0.42044736 100      806     1
#> 807   0.0449060511 0.34028748 100      807     1
#> 808   0.0180786956 1.81800598 100      808     1
#> 809  -1.5087824116 2.30128835 100      809     1
#> 810   2.8326917406 0.89591163 100      810     1
#> 811   4.8439433967 2.31049521 100      811     1
#> 812  -2.0510214421 2.27381132 100      812     1
#> 813  -1.3027921410 0.70100852 100      813     1
#> 814  -1.1672767337 3.30513907 100      814     1
#> 815  -0.5959152818 1.87281971 100      815     1
#> 816   0.4972307957 0.58679972 100      816     1
#> 817  -3.7347997094 2.55235292 100      817     1
#> 818   0.1542379212 0.07206258 100      818     1
#> 819   0.6310278549 0.94917253 100      819     1
#> 820   1.2731890907 2.64745954 100      820     1
#> 821   0.5263316794 1.35463058 100      821     1
#> 822  -0.0149303598 0.02384703 100      822     1
#> 823   0.2673043672 0.52814542 100      823     1
#> 824   1.8430668079 2.44556519 100      824     1
#> 825  -0.3150833651 0.30975079 100      825     1
#> 826   0.4811093097 1.76310730 100      826     1
#> 827   0.5424986098 0.85330577 100      827     1
#> 828  -4.1237945958 2.52098717 100      828     1
#> 829  -2.0728093036 1.44080344 100      829     1
#> 830   0.3969593711 0.60035165 100      830     1
#> 831  -0.2444911530 1.65850453 100      831     1
#> 832  -1.2886043403 1.38910362 100      832     1
#> 833   0.5395160226 0.67038175 100      833     1
#> 834  -0.0723965029 1.51582780 100      834     1
#> 835   7.3162517505 2.96937847 100      835     1
#> 836  -0.2645356177 2.04880466 100      836     1
#> 837  -0.5979090267 1.44478691 100      837     1
#> 838  -0.4018340062 0.39409556 100      838     1
#> 839  -0.0158119272 1.26086406 100      839     1
#> 840   1.0133583283 0.87781073 100      840     1
#> 841  -0.2857150694 0.34239986 100      841     1
#> 842   1.0847281154 1.96458248 100      842     1
#> 843  -5.1104246995 2.13511912 100      843     1
#> 844   0.8137394273 2.86395872 100      844     1
#> 845   0.7416322773 0.70690097 100      845     1
#> 846  -0.8722044380 1.33232884 100      846     1
#> 847   0.0167750878 0.64462945 100      847     1
#> 848   0.1834665640 0.23219012 100      848     1
#> 849   1.6201699882 3.44938884 100      849     1
#> 850   0.0226004655 0.14317184 100      850     1
#> 851   0.1967064070 0.46297580 100      851     1
#> 852  -0.3313009552 0.41518488 100      852     1
#> 853  -0.2998668672 0.59997310 100      853     1
#> 854  -1.5258617254 0.69713330 100      854     1
#> 855   1.4483751577 1.46091174 100      855     1
#> 856   0.4016254005 0.49580724 100      856     1
#> 857   6.5455412842 2.99824394 100      857     1
#> 858   0.4641334842 0.95121104 100      858     1
#> 859   0.8106038096 1.23455815 100      859     1
#> 860   0.0443618274 0.13337275 100      860     1
#> 861   0.4303865856 0.30969904 100      861     1
#> 862  -0.4745598644 0.45841239 100      862     1
#> 863  -0.2494850571 0.77078200 100      863     1
#> 864   2.1214110646 1.75389179 100      864     1
#> 865  -0.1174320041 0.30681698 100      865     1
#> 866  -3.1480281028 1.78036075 100      866     1
#> 867  -3.2955107980 2.22838322 100      867     1
#> 868   2.6126012062 3.08712638 100      868     1
#> 869  -1.0766397785 1.46773221 100      869     1
#> 870   1.1455183656 2.56388989 100      870     1
#> 871  -1.7168923696 0.72539006 100      871     1
#> 872  -2.3865455439 2.90157831 100      872     1
#> 873   1.3677838875 1.39732708 100      873     1
#> 874  -0.0218301203 0.79748598 100      874     1
#> 875  -1.6566584325 2.82960482 100      875     1
#> 876  -0.6703879229 2.64764944 100      876     1
#> 877  -2.2417086552 1.61198996 100      877     1
#> 878  -0.6650041411 0.82054932 100      878     1
#> 879  -2.3484965187 2.21479129 100      879     1
#> 880   0.0825926731 0.11296314 100      880     1
#> 881   0.1987879087 0.79508923 100      881     1
#> 882  -1.1935488093 1.40792849 100      882     1
#> 883  -0.1116625167 2.02221133 100      883     1
#> 884  -0.1463031471 1.28481807 100      884     1
#> 885   0.4705192636 0.46194147 100      885     1
#> 886   0.3504712347 2.59647924 100      886     1
#> 887  -0.6797155983 1.03076533 100      887     1
#> 888  -1.6831485716 2.77582025 100      888     1
#> 889  -0.1165407464 1.49778183 100      889     1
#> 890  -0.8647843117 1.18681132 100      890     1
#> 891  -2.9932393458 1.22687499 100      891     1
#> 892  -0.4976446819 0.89642862 100      892     1
#> 893  -0.6383177093 1.71366473 100      893     1
#> 894  -1.2631307069 1.47548269 100      894     1
#> 895  -0.9484269772 2.45975067 100      895     1
#> 896   0.6560466461 1.49881434 100      896     1
#> 897   0.9865365823 0.85957949 100      897     1
#> 898   0.2183178612 1.55721599 100      898     1
#> 899  -0.0910747707 0.93093417 100      899     1
#> 900  -1.9366222803 1.94646082 100      900     1
#> 901   2.2398895096 2.97515221 100      901     1
#> 902  -1.4384148778 1.84085033 100      902     1
#> 903  -1.5030903535 1.56459663 100      903     1
#> 904   0.8059545459 1.25893717 100      904     1
#> 905   0.3095326005 0.47358046 100      905     1
#> 906   0.8801656881 1.45629181 100      906     1
#> 907  -0.2214232072 1.00841295 100      907     1
#> 908   0.9013883772 1.57002746 100      908     1
#> 909   0.6842475958 0.34317117 100      909     1
#> 910   0.1481848709 0.22513650 100      910     1
#> 911   1.5801877589 1.92073437 100      911     1
#> 912   0.1025105501 1.46628121 100      912     1
#> 913  -0.4549931399 0.79814441 100      913     1
#> 914   0.2321461152 2.02166237 100      914     1
#> 915   0.1357008873 0.36157567 100      915     1
#> 916  -1.1505449928 2.55422162 100      916     1
#> 917   0.6303535500 2.46865050 100      917     1
#> 918   1.3226180522 2.28933072 100      918     1
#> 919  -0.1762531415 1.27342955 100      919     1
#> 920  -1.3008539465 0.99024942 100      920     1
#> 921   1.5609005452 2.22940244 100      921     1
#> 922   0.5864884747 0.84371766 100      922     1
#> 923   2.3581202517 3.00427309 100      923     1
#> 924  -1.9621666737 1.12400231 100      924     1
#> 925   3.6213140510 2.02329056 100      925     1
#> 926  -4.1091430547 2.17960989 100      926     1
#> 927  -0.5061457878 0.64117486 100      927     1
#> 928  -1.2684881474 1.69475362 100      928     1
#> 929  -0.1851318564 2.36169493 100      929     1
#> 930  -0.7374841628 1.11142565 100      930     1
#> 931  -0.4258743934 0.22096350 100      931     1
#> 932  -1.9031377165 1.31311427 100      932     1
#> 933   0.5276163912 1.56893373 100      933     1
#> 934   0.9398751119 0.96351410 100      934     1
#> 935  -0.0860299478 1.51926373 100      935     1
#> 936   0.3804690727 0.16979182 100      936     1
#> 937  -0.1712302662 1.73484652 100      937     1
#> 938   0.5628675457 1.01117294 100      938     1
#> 939   1.1268649255 1.30849435 100      939     1
#> 940   1.3380973053 3.14105936 100      940     1
#> 941  -3.6431499813 1.91797128 100      941     1
#> 942  -1.2217484127 1.48571472 100      942     1
#> 943  -0.3279084110 0.70703462 100      943     1
#> 944  -0.1293770565 0.53471004 100      944     1
#> 945   1.2512452156 0.69743750 100      945     1
#> 946  -0.2383829521 0.79237153 100      946     1
#> 947   0.0685775545 0.03268754 100      947     1
#> 948  -0.0873384517 2.29257704 100      948     1
#> 949   0.5873215145 0.81480659 100      949     1
#> 950   0.4683596374 1.21906284 100      950     1
#> 951   0.3651318252 1.84988104 100      951     1
#> 952  -1.1734694172 0.90576084 100      952     1
#> 953   0.0191623121 1.69928380 100      953     1
#> 954  -1.3504800168 2.54109400 100      954     1
#> 955   0.7469353297 0.54083861 100      955     1
#> 956   0.3028922108 1.60304701 100      956     1
#> 957   3.1306283864 2.80364046 100      957     1
#> 958   4.1064663158 2.01293514 100      958     1
#> 959  -0.1063672130 2.19740946 100      959     1
#> 960   0.5406606619 1.34715842 100      960     1
#> 961   0.0380851090 0.12799292 100      961     1
#> 962  -0.3399440121 1.33641880 100      962     1
#> 963   0.9901341520 2.03897306 100      963     1
#> 964  -0.5122165068 0.51408720 100      964     1
#> 965   0.2064450874 0.12704981 100      965     1
#> 966   3.1167894554 3.12248314 100      966     1
#> 967   0.8106375213 0.75931601 100      967     1
#> 968   1.4396072810 1.17898386 100      968     1
#> 969  -0.2730442503 1.33037244 100      969     1
#> 970  -0.0411198669 0.04445097 100      970     1
#> 971  -0.8815240727 2.15955069 100      971     1
#> 972  -2.5070371194 1.86418083 100      972     1
#> 973  -0.2158038176 0.85654243 100      973     1
#> 974   0.6850490441 1.47189046 100      974     1
#> 975  -1.3922933190 1.26588961 100      975     1
#> 976  -1.4385978425 1.09375335 100      976     1
#> 977   0.0907377417 0.09269724 100      977     1
#> 978  -0.4384580956 0.42370080 100      978     1
#> 979   0.0960999474 0.27455960 100      979     1
#> 980  -0.4181972744 0.90124163 100      980     1
#> 981  -2.2216589926 2.73410629 100      981     1
#> 982   3.3271731488 2.43502355 100      982     1
#> 983   0.3271672321 0.44653967 100      983     1
#> 984  -1.3035047099 1.43275417 100      984     1
#> 985   0.2025504527 0.84458900 100      985     1
#> 986  -0.1352997388 1.68416056 100      986     1
#> 987   1.9046454456 1.73113958 100      987     1
#> 988  -0.8213056217 1.02087601 100      988     1
#> 989   1.2925963994 1.33093189 100      989     1
#> 990  -0.4431189902 1.97181659 100      990     1
#> 991  -0.4683925149 0.79821492 100      991     1
#> 992  -1.6328162592 1.39467525 100      992     1
#> 993  -1.6693517639 2.68397373 100      993     1
#> 994   0.0079014258 0.29581427 100      994     1
#> 995  -0.0176643595 0.02328736 100      995     1
#> 996  -1.1364467895 3.33054381 100      996     1
#> 997   1.0358121840 2.32782954 100      997     1
#> 998   0.6860441027 1.97487240 100      998     1
#> 999  -0.1618695875 2.24892675 100      999     1
#> 1000 -4.1867825265 3.19446065 100     1000     1

simulate_dgm("Stanley2017", list(environment = "Cohens_d", mean_effect = 0,
                        effect_heterogeneity = 0, bias = 0, n_studies = 5,
                        sample_sizes = c(32,64,125,250,500)))
#>             yi        sei   ni
#> 1 -0.219640178 0.25075265   64
#> 2 -0.089021210 0.17686423  128
#> 3 -0.244062509 0.12696115  250
#> 4 -0.110875501 0.08951141  500
#> 5  0.008493387 0.06324584 1000