Upload Datasets of a DGM

This function uploads datasets of a specified Data-Generating Mechanism (DGM) to the OSF repository at https://osf.io/exf3m/.

This is an internal function intended for the benchmark maintainer. It requires OSF repository authentication (via osfr::osf_auth()) and repository access.

Usage

upload_dgm_datasets(dgm_name, path = NULL, overwrite = TRUE, progress = TRUE)

upload_dgm_results(dgm_name, path = NULL, overwrite = TRUE, progress = TRUE)

upload_dgm_measures(dgm_name, path = NULL, overwrite = TRUE, progress = TRUE)

Arguments

dgm_name

Character string specifying the name of the DGM dataset to upload.

path

Character string specifying the directory path where the datasets/results/measures are located. Defaults to the location specified via PublicationBiasBenchmark.get_option("simulation_directory"). The objects are read from dgm_name/datasets, dgm_name/results, dgm_name/measures subfolders.

overwrite

Logical indicating whether to overwrite existing files on OSF. Defaults to TRUE, which overwrites files with conflicts.

progress

Logical indicating whether to show progress uploading files. Defaults to TRUE.

Value

TRUE if the upload was successful, otherwise an error is raised.

Examples

if (FALSE) { # \dontrun{
  upload_dgm_datasets("no_bias")
} # }